% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{auc}
\alias{auc}
\title{Returns the area under the ROC curve governed by the input parameter.}
\usage{
auc(theta)
}
\arguments{
\item{theta}{The parameter.}
}
\value{
The area under the curve.
}
\description{
Returns the area under the ROC curve governed by the input parameter.
}
