% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{lehmann_roc}
\alias{lehmann_roc}
\title{Lehmann model for ROC curves.}
\usage{
lehmann_roc(formula, data, cov_vals = NULL)
}
\arguments{
\item{formula}{A formula containing the covariates to be analyzed.}

\item{data}{A data.frame containing the variables named in the formula.}

\item{cov_vals}{An optional list of covariate values.}
}
\value{
The output is a list of relavant statistics. \code{theta},
\code{var_theta}, \code{auc}, and \code{var_auc} are all numerics.
\code{partial_auc} and \code{var_partial_auc} are both functions of the
false positive rate. \code{roc} is a list containing numeric vectors
\code{TPR} and \code{FPR}. \code{var_roc} is a function of the false
positive rate. If one or more covariates are named in the formula, all
numeric values in the return list become functions of the covariates.
If a list of covariates, \code{cov_vals}, is provided, the covariate values
will be passed into the functions and the output will again contain
numerics.
}
\description{
Relevant statistics for a Lehmann model of an ROC curve.
}
\examples{
# Import the ovarian dataset
ovarian <- survival::ovarian
head(ovarian)

# Create a Lehmann ROC model using only disease status.
l <- lehmann_roc(futime~resid.ds, ovarian)
summary(l)
plot(l)
l$theta
l$auc

# Create a Lehmann ROC model with age as a concomitant covariate.
l <- lehmann_roc(futime~resid.ds*age, ovarian)

# The output is now a list of functions
summary(l)

theta <- l$theta
theta_at_60 <- theta(list(age=60))
theta_at_60
l$auc(theta_at_60)
roc <- l$roc(theta_at_60)
plot(roc$FPR, roc$TPR)

# Run the same model but include a list of covariate values.
l <- lehmann_roc(futime~resid.ds*age, ovarian, list(age=60))
summary(l)
plot(l)
l$theta
l$auc
}
