% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{apply_covariates}
\alias{apply_covariates}
\title{Returns a data.frame containing values of \code{theta}, \code{var_theta}
\code{auc}, and \code{var_auc}}
\usage{
apply_covariates(l, covariates, ...)
}
\arguments{
\item{l}{A lehmann_roc object.}

\item{covariates}{A list of covariate values}

\item{...}{More covariate values}
}
\value{
A data.frame containing the values of relavant statistics
with the covariate values they were computed from.
}
\description{
This function is used when a user wants apply covariate values
}
