% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc.R
\name{lehmann_roc}
\alias{lehmann_roc}
\title{Lehmann model for ROC curves.}
\usage{
lehmann_roc(formula, data)
}
\arguments{
\item{formula}{A formula containing the covariates to be analyzed.}

\item{data}{A data.frame containing the variables named in the formula.}
}
\value{
The output is a list of relavant statistics. \code{theta},
\code{var_theta}, \code{auc}, and \code{var_auc} are all numerics.
\code{partial_auc} and \code{var_partial_auc} are both functions of the
false positive rate. \code{roc} is a list containing numeric vectors
\code{TPR} and \code{FPR}. \code{var_roc} is a function of the false
positive rate. If the input formula contains concomitant covariates,
the return list will contain functions. For example, \code{theta}
would now be a function of the covariate values. The \code{apply_covariates}
function should be used to input values of covariates into appropriate
lehmann_roc objects.
}
\description{
Relevant statistics for Lehmann models of ROC curves.
}
\examples{
# Import a dataset
ovarian <- survival::ovarian
head(ovarian)

# Create a Lehmann ROC model using only disease status.
l <- lehmann_roc(futime~resid.ds, ovarian)
summary(l)
plot(l)
l$theta
l$auc
l$partial_auc(0.5)

# Create a Lehmann ROC model with age as a concomitant covariate.
l <- lehmann_roc(futime~resid.ds*age, ovarian)

# The output is now a list of functions
summary(l)

\dontrun{
l # errors
plot(l) # errors
}

# Use the apply_covariates function
l_objects <- apply_covariates(l, list(age=21))
l2 <- l_objects[["21"]]
l2
l2$partial_auc(0.5)
plot(l2)

}
